pyproject.toml
[project]
name = "namecheap-connector"
version = "0.1.0"
description = "A utility for connecting to Namecheap DNS registrar"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "flake8>=7.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "structlog>=25.4.0",
]
